:recycle:  Refactor: import 경로 절대경로로 수정, dummy data 경로 변경
Closes #135
