--- SampleData/RadialChartComponent.js
@@ -1,50 +1,0 @@
-export const sampleData = [
-  {
-    categoryId: 1,
-    name: 'react',
-    averageScore: 67,
-    colorCode: '#333',
-    postCount: 86,
-    isPublic: true,
-  },
-  {
-    categoryId: 2,
-    name: 'react',
-    averageScore: 76,
-    colorCode: '#333',
-    postCount: 75,
-    isPublic: true,
-  },
-  {
-    categoryId: 3,
-    name: 'vue',
-    averageScore: 63,
-    colorCode: '#333',
-    postCount: 90,
-    isPublic: true,
-  },
-  {
-    categoryId: 4,
-    name: 'angular',
-    averageScore: 89,
-    colorCode: '#333',
-    postCount: 76,
-    isPublic: true,
-  },
-  {
-    categoryId: 5,
-    name: 'spring',
-    averageScore: 96,
-    colorCode: '#333',
-    postCount: 97,
-    isPublic: true,
-  },
-  {
-    categoryId: 6,
-    name: 'java',
-    averageScore: 67,
-    colorCode: '#333',
-    postCount: 85,
-    isPublic: true,
-  },
-]
